Add delta measurement and add a RigidObject class

+++ headers/Object.hpp
@@ -8,7 +8,7 @@ class Object
 public:
 	Object();
 
-	virtual void move(int x, int y);
+	virtual void move(float x, float y);
 	virtual void update(uint32_t delta);
 	virtual void paint(SDL_Surface* surface);
